// src/interfaces/spotify.ts
interface AlbumObjectFull extends AlbumObjectSimplified {
    artists: ArtistObjectSimplified[];
    copyrights: CopyrightObject[];
    external_ids: ExternalIdObject;
    genres: string[];
    popularity: number;
    release_date: string;
    release_date_precision: string;
    tracks: PagingObject<TrackObjectSimplified>;
}

/**
 * Simplified Album Object
 * [album object (simplified)](https://developer.spotify.com/web-api/object-model/#album-object-simplified)
 */
interface AlbumObjectSimplified {
    album_type: string;
    available_markets?: string[];
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    images: ImageObject[];
    name: string;
    type: 'album';
    uri: string;
    [rest:string]:any;
}

/**
 * Full Artist Object
 * [artist object (full)](https://developer.spotify.com/web-api/object-model/)
 */
interface ArtistObjectFull extends ArtistObjectSimplified {
    followers: FollowersObject;
    genres: string[];
    images: ImageObject[];
    popularity: number;
}

/**
 * Simplified Artist Object
 * [artist object (simplified)](https://developer.spotify.com/web-api/object-model/)
 */
export interface ArtistObjectSimplified {
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    name: string;
    type: 'artist';
    uri: string;
}
// ...
interface CopyrightObject {
    text: string;
    type: 'C' | 'P';
}
// ...
interface ExternalIdObject {
    isrc?: string;
    ean?: string;
    upc?: string;
}

/**
 * External Url Object
 * [](https://developer.spotify.com/web-api/object-model/)
 *
 * Note that there might be other types available, it couldn't be found in the docs.
 */
interface ExternalUrlObject {
    spotify: string;
}

/**
 * Followers Object
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface FollowersObject {
    href: string;
    total: number;
}

/**
 * Image Object
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface ImageObject {
    height?: number;
    url: string;
    width?: number;
}

/**
 * Paging Object wrapper used for retrieving collections from the Spotify API.
 * [](https://developer.spotify.com/web-api/object-model/#paging-object)
 */
export interface PagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string;
    offset: number;
    previous: string;
    total: number;
}
// ...
interface PlaylistBaseObject {
    collaborative: boolean;
    description: string | null;
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    images: ImageObject[];
    name: string;
    owner: UserObjectPublic;
    public: boolean;
    snapshot_id: string;
    type: 'playlist';
    uri: string;
}

/**
 * Playlist Object Full
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface PlaylistObjectFull extends PlaylistBaseObject {
    followers: FollowersObject;
    tracks: PagingObject<PlaylistTrackObject>;
}

/**
 * Playlist Object Simplified
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface PlaylistObjectSimplified extends PlaylistBaseObject {
    tracks: {
        href: string;
        total: number;
    };
}
// ...
export interface SavedAlbumObject {
    added_at: string;
    album: AlbumObjectFull;
}
// ...
interface TrackObjectSimplified {
    artists: ArtistObjectSimplified[];
    available_markets?: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    is_playable?: boolean;
    linked_from?: TrackLinkObject;
    name: string;
    preview_url: string;
    track_number: number;
    type: 'track';
    uri: string;
}
// ...
interface UserObjectPublic {
    display_name?: string;
    external_urls: ExternalUrlObject;
    followers?: FollowersObject;
    href: string;
    id: string;
    images?: ImageObject[];
    type: 'user';
    uri: string;
}